Planning the Update Statement
This patch takes the result of the analysis of the UPDATE statement and
generates the necessary query plan. The query plan builds consists of
two nodes: a scan node, that will generate the rows with the values and
key columns that will be updated and a table sink that will process the
incoming row batches and send them to Kudu. Since there are no multi-row
transactions, an error aborting the update process in the middle will
leave some of the rows updated and other in their original state.

Change-Id: I30d62eb0074a5ac3c1cff1058c00485da02c30a3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6998
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: jenkins