IMPALA-2062: Replace DCHECK_NOTNULL(X) with DCHECK(X != NULL)
Google's headers define DCHECK_NOTNULL so that even in release builds
the argument is evaluated. This is different from the usual DCHECK()
macros. We redefined DCHECK_NOTNULL to be a no-op in release builds, but
instead we should just be using DCHECK(X != NULL). One reason is that as
we import code from other sources that had used these macros, they would
reasonably expect the macro semantics to be the same, leading to subtle
bugs that only manifest in release builds.

Change-Id: Id8c08faea3099d52cd5604209384a7cca79160a1
Reviewed-on: http://gerrit.cloudera.org:8080/486
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: Internal Jenkins