IMPALA-1702: Fix crash when IMPALA-1702 hit even with workaround patch
IMPALA-1702 results in bad partition indexes. This would cause Impala to crash as the
partition descriptor object for that index would be NULL, and Impala would happily try to
dereference it.

To provide a workaround, we added a patch which would not DCHECK, but would gracefully
return when this condition was detected. However, the graceful exit wasn't extended to
HdfsScanNode::Close(), which is called after Prepare() even if Prepare() failed.

This patch fixes the issue by ensuring that Close() also gracefully
fails.

To defensively protect against the situation where it may not be safe
to Close() an ExprContext that hasn't been Prepare()'d, this patch adds
a guard to HdfsPartitionDescriptor::CloseExprs() to ensure that only
Prepare()'d exprs are Close()'d. This is necessary in this patch for the
same reasons as the crash above: the scan node may Close() after an
unsuccessful Prepare().

IMPALA-2156 is related, as some of this defensive code would be easier
if the lifecycle of exprs was better specified.

Testing was manual, by editing the partition metadata to inject bad partition indexes.

Change-Id: I3d5a20871ab83dfa2d6aae85ca6610ddee93daac
Reviewed-on: http://gerrit.cloudera.org:8080/534
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 154830d919bc972ccad49628ab2c6d99addc8dfd)
(cherry picked from commit 05b978d27b5ea1ce9fd336c763d2920d36781395)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/11710
Reviewed-by: Bharath Vissapragada <bharathv@cloudera.com>
Tested-by: Bharath Vissapragada <bharathv@cloudera.com>