IMPALA-2141: UnionNode::GetNext() doesn't check for query errors
When a UDF with constant parameters in the select list calls SetError(), it does not fail
the query. This is because UnionNode::GetNext() does not check for errors after
UnionNode::EvalAndMaterializeExprs() evaluates the expression, which itself does not
report the error.

Change-Id: I8850cf1a603e320bb23f4a9a4d47600d14590f3a
(cherry picked from commit 011bd4cbbc9a9f17fc2c803e331a573cb610b2d2)