IMPALA-2081: Add PERCENT_RANK, NTILE, CUME_DIST analytic window functions
These functions are implemented as rewrites in the analysis stage. They are rewritten as
different arithmetic expressions and make use of the existing analytic functions such as
'rank', 'count' and 'row_number' to compute the final results.

TODO: IMPALA-2171: NTILE() currently takes only constant expressions. We need to modify
it to take non-constant expressions as well in a future patch.

Change-Id: I8773df8ceefff27ab66a41169dc4ac0927465191
Reviewed-on: http://gerrit.cloudera.org:8080/584
Tested-by: Internal Jenkins
Reviewed-by: Henry Robinson <henry@cloudera.com>