IMPALA-2165: Avoid cardinality 0 in scan nodes of small tables and low selectivity
To estimate the cardinality of the scan nodes we multiply the number
of rows of the table with the computed selectivity (which is a
function of the number of predicates). If the table is small (e.g. a
small dimension table) and there are a few predicates, then because of
rounding we end up using cardinality of 0 for this scan node. But this
is causing errors in the cardinality estimation of upstream nodes such
as cross join that typically multiply the cardinalities of their
inputs.

This patch fixes the problem by making sure that in case we round to
0 the cardinality, we use 1 instead. Some plans change because of this
patch.

Change-Id: Ic52e4f40964a5620f674803b48c4af814e6d4164
Reviewed-on: http://gerrit.cloudera.org:8080/819
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 29ffd7927aaa0e2ea4a9b21c7a88ee3a3b8a6897)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8708
Reviewed-by: Juan Yu <jyu@cloudera.com>
Tested-by: Juan Yu <jyu@cloudera.com>
(cherry picked from commit facf3d8034d31f0504d3c8d71072b7dcbf8a2b1f)