IMPALA-2201: Unconditionally update the partition stats and row count.
Before this patch, we used to only send alterPartition() requests to
the Hive Metastore for partitions whose stats have changed during
COMPUTE [INCREMENTAL] STATS. However, there is other state associated
with the stats like the STATS_GENERATED_VIA_STATS_TASK that was not
properly handled. Not updating this additional partition metadata
was the root cause of IMPALA-2201.

This patch changes COMPUTE [INCREMENTAL] STATS to unconditionally update the
partition stats and row counts in the Hive Metastore, even if the partition
already has identical stats. This behavior results in possibly redundant work,
but it is predictable and easy to reason about because it does not depend on
the existing state of the metadata.

Note that in versions starting from CDH 5.4 it is not possible to reproduce
IMPALA-2201 because of a behavioral change in the Hive Metastore in the
alterPartition() code path.

Change-Id: I10105d8d6306d9ad9988b03abc23752d7bc98252
Reviewed-on: http://gerrit.cloudera.org:8080/640
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins