KUDU-1010 - Pass a new instance of KuduPredicate to each new KuduScanner
KuduScanner takes ownership of the KuduPredicates that are passed
in AddConjunctPredicate, however, since we're building the predicates
only once in KuduScanNode, we would crash with a SIGSEGV when a new scanner
is created.

This patch changes impala's KuduScanner to take advantage of the new Clone()
method and clone the predicates that are passed to kudu's KuduScanner.
This also addresses a long standing TODO in kudu-scan-node-test that mentioned
we were missing tests that scan multiple tablets. Now each test must
specify the ranges that should be scanned.

Change-Id: I693d204e6008e76e0fa76b455ce820e9f67ae936
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7497
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Tested-by: jenkins