IMPALA-2075: Report the correct time spent in the Join Node
The definition of "local time" is the wall clock time where this exec
node is processing and it is not blocked by any of its children.

The join node has two execution models:
  1. The entire join execution is in a single thread.
  2. The build side is executed on a different thread while the main
     thread opens the probe (left) child.

In case 1, the "local time" spent in this node is as simple as:
    total_time - left child time - right child time
Because the entire right child time blocks the execution, the right
child time is the same as right_child_blocking_stop_watch_.

Case 2 is more complicated. The build thread is started first and
then the main thread will "open" the left child. When the left child
is ready (i.e. Open() returned), the main thread will wait for the
build thread to finish. Because the left child is always executed in
the main thread, all the left child time should not be counted
towards the hash join "local time".
For the right child (the build side), the child time in the build
thread up to the point when the left child Open() returns should not
be counted towards the hash join local time. This time period
completely overlaps with the left child time. From the time when left
child Open() returned, the right child time should be removed from
the total time because this is the only child that is blocking the
join execution.

Here's the calculation:
  total_time - left child time - (right child time - overlapped
  period)

The "overlapped period" is measured by a stop watch. Using this
overlap method, both children's "Prepare" time are also excluded.

The "local time" for the hash join node is now reported by a "derived counter"
using the above equation.

Change-Id: Icd518ca8e346836f010f663a2a1713b7a41ad0fc
Reviewed-on: http://gerrit.cloudera.org:8080/1170
Reviewed-by: Casey Ching <casey@cloudera.com>
Tested-by: Internal Jenkins