KUDU-1093: Lazily relocate string values into row mem pool
While the row batch is statically allocated in the scanner, the
auxiliary memory for strings is required dynamically. To avoid copying
data from the KuduRowResult into intermediate structures before copying
into the row batch, this patch will only copy the pointer to the
location in the KuduRowResult. Once the conjuncts have been evaluated
the string values are materialized as well.

This reduces the memory pressure for highly selective queries, in
addition it avoids copying the string data for those rows which don't
pass the conjuncts.

Change-Id: Ib5456a4e775fef1508b6e0552c31e89419209c04
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7772
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins