Python: Switch query generator back to psycopg2 driver
The pg8000 driver currently in use doesn't work well with autocommit
enabled. When enabled, the drive complains that only 100 rows can be
fetched because of a buffer/cache limitation. With autocommit disabled,
a bunch of code changes would be needed. For now the driver the
previous psycopg2 driver will be used again. (The psycopg2 driver was
originally replaced because dev postgres libraries are required to
build it, so building the virtualenv would fail. This patch doesn't
try to build psycopg2 and instead assumes it was already installed.)

Change-Id: I6901cb1fa109d6da907b1415601116d833d66656
Reviewed-on: http://gerrit.cloudera.org:8080/737
Reviewed-by: Ishaan Joshi <ishaan@cloudera.com>
Tested-by: Casey Ching <casey@cloudera.com>