IMPALA-2283: Occasional crash in thrift-server-test
A running ThriftServer cannot be destroyed safely. A race condition
inside Thrift may cause some object to be referenced during its own
destruction, which can lead to a pure virtual call invocation when the
object has been reduced to an abstract base class. This most likely
happens when TThreadedServer wakes up from recv().

Rather than try to fix the underlying issue, this patch sidesteps it by
allocating ThriftServer objects on the heap, and failing to delete
them. Since the d'tor is never called, the race condition is never
triggered, and since this is only an issue in test code the memory leak
is not significant.

Testing:

Before this patch, the race would manifest itself once every ~10
minutes. After this patch, the tests have run in a loop for 30 minutes
with no errors.

Change-Id: I5be3bb2ec013053a8cd4e9341c162833861c75fa
Reviewed-on: http://gerrit.cloudera.org:8080/746
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins