KUDU-996: Make DDL ops and scans case independent.
In general SQL is (typically) case-agnostic and Impala normalizes all
columns to the lower case version when querying. This patch makes sure,
that key columns get persisted as lower case and that the table
properties for the key columns that are parsed when loading a table are
treated as lower case.

Change-Id: I085d8c67c2cd4ded335ee8257f73332c19429fbe
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7819
Reviewed-by: Alexander Behm <alex.behm@cloudera.com>
Tested-by: jenkins