IMPALA-2278: fix generate-schema-statements-py --force-reload for text/lzo
The combination of --force and text/lzo was broken if the partition
directories already contained data.  For reasons explained in the
comments, the ALTER TABLE ADD PARTITION step is skipped in this case,
which causes HIVE to not do a full overwrite with INSERT OVERWRITE.

Fix it by manually removing the directories.

Testing: Verified the following combinations of load-data.py for
	 text/lzo now work:
 {--force, ""} x {no partition dirs, partition dirs with files}

Change-Id: I3ee34c4d85c58644345eadd8fc0976665c1bbaf5
Reviewed-on: http://gerrit.cloudera.org:8080/752
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Tested-by: Internal Jenkins