Fix race condition in test_statestore changes
KillableThreadedServer.port is required to be set for
KillableThreadedServer.wait_until_up() to execute. However, it was being
set during serve() which was run concurrently with wait_until_up() (see
StatestoreSubscriber.__init_server()), so sometimes it was not set. The
fix is to set KillableThreadedServer.port during construction (it is set
by the underlying socket as soon as it is constructed itself).

Change-Id: Ib9ca9e237bca96635f5ee5b5bbfb7fd678929ce4
Reviewed-on: http://gerrit.cloudera.org:8080/759
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
Tested-by: Internal Jenkins