Fix race in DiskIoMgr.
I discovered this while reading through the code to help with an Impala issue. The
problem is that in the cancellation path, EnqueueBuffer() will return the buffer
descriptor to the cache, which is effectively the same as freeing it. The caller
would then try to read some state from it and the state can be corrupt. From a conceptual
point of view, EnqueueBuffer() represents a memory hand-off so this behavior makes
sense. The caller just needs to store the state it needs before EnqueueBuffer().

This fix is similar to that for IMPALA-2101.

Change-Id: I2deab4923d1324fd275b8dd4ad37eedb6d39922c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8023
Tested-by: jenkins
Reviewed-by: Nong Li <nong@cloudera.com>
Reviewed-on: http://gerrit.cloudera.org:8080/834
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: Internal Jenkins