KUDU-1144 Build predicate KuduValues based on the column type, not on the expression literal type
The frontend sends numeric literals to the backend either as an int
literal or as a float literal. KuduValue has a single builder method
for ints, which casts to int64_t, so int literals work. However, KuduValue
distinguishes between double and float but the impala frontend always
sends a float literal. This means that, if the predicate is over a
double column, we would build a KuduValue with FromFloat(), which is wrong
and would be rejected by the scanner. Because we were also ignoring the Status
on scanner_->AddConjunctPredicate() the predicate would get completely ignored,
making the query return wrong results.

This makes it so that scan node builds KuduValues that are set on
predicates based on the type of the column and not on the type of
the expression literal.

This also makes it so that we check the returned status
scanner_->SetConjunctPredicate() which we were ignoring before and
was the cause of KUDU-1144.

It's hard to test this without violating what the frontend would send/push down.
I tried: 'select count(*) from lineitem_sf1000 where l_discount >= 100'
and it returned the correct result (0) versus returning all rows, as
before, as it was the predicate was being ignored.

Change-Id: I6c53e1ea2f00c2c28cd100b1c83a2a66ba2258e2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8061
Reviewed-by: Martin Grund <mgrund@cloudera.com>
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins