IMPALA-2314: LargestSpilledPartition was not checking if partition is closed
In both PHJ and PAGG we call LargestSpilledPartition() in order to
figure out if there was any reduction in the size of the partitions
after a repartition. This code had a mistake as it was not checking
first whether a partition has been closed, which could happen if, for
example, a partition ended up empty after repartitioning. The result
of that it was SEGV as we were trying to reference a stream that had
already been set to NULL.

Change-Id: Ia369b0b3ad19c62da05e7ed1bcc4d28ebbe8c9df
Reviewed-on: http://gerrit.cloudera.org:8080/875
Reviewed-by: Ippokratis Pandis <ipandis@cloudera.com>
Tested-by: Internal Jenkins