Fix casing issues in scanner predicate pushdown
We fixed casing problems when setting projections on the Kudu scanner, but didn't
on predicate pushdown. This meaning that predicates whose slot refs referred to Kudu
columns with a different casing would cause the query to fail.

This patch fixes that and adds where clause to an existing test with casing mismatch
to make sure it now works, where before we would get an error.

Change-Id: I686744a1fd574282290a5f5c7cf01f00e73c9411
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/8441
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins