IMPALA-2444: Part 1: Parquet scanner perf improvements
This is a set of related changes that avoid executing nested
types-specific logic in cases when it is not needed. The additional
branches had a measurable performance overhead.

Add template arguments for AssembleRows and ReadRow so that code is
separately optimized for the case when we're assembling collection rows
versus top-level rows. Several branches are not needed when assembling
top-level rows.

Reorganise ColumnReader so that ReadValue is the virtual function rather
than ReadSlot, and so that ReadValue is called even if we are not
materializing slots. This lets use add a template argument to
ScalarColumnReader so we can produce a specialized class to handle
non-materialized slots, and so we can avoid checking whether it has a
slot descriptor in ReadRow. This also lets us explicitly specify which
version of NextLevels is being called to guarantee it won't be called
virtually.

This brings the regression down to ~15% down from ~20%.

Change-Id: I494f1c8e139e6d39c0fc93a2ca841f59ff867717
Reviewed-on: http://gerrit.cloudera.org:8080/1088
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins