IMPALA-2441: CREATE DATABASE IF NOT EXISTS may cause NPE
This commit fixes an issue where a CREATE DATABASE IF NOT EXISTS
statement will cause an NPE if the database already exists in the
Hive MetaStore but not in the Impala catalog. With this fix no error is
thrown if the database exists in HMS and the database is added to the
catalog.

Change-Id: If1d15bb50869ce8084e0443f119a596b365004c7
Reviewed-on: http://gerrit.cloudera.org:8080/1091
Reviewed-by: Dimitris Tsirogiannis <dtsirogiannis@cloudera.com>
Tested-by: Internal Jenkins