IMPALA-2444: Part 5: Parquet scanner perf improvements.
This patch is a mix of microoptimisations.

One set of microoptimisations avoid branches and instruction cache pollution
in the decoders. This buys a couple of % improvement.

The other set of microoptimisations makes the Parquet scanner more
cache efficient by moving frequently accessed members towards the
beginning of the class, inlining frequently accessed variables,
and by compacting members. These significantly improve perf:
after this patch the Parquet scanner is 4-5% faster than in Impala
2.4.7.

Levels are represented with int16_t, which is sufficient since the
decoders produce uint8_ts.  We can then include max_rep_level_ and
max_def_level_ inline in the struct without using additional space.

Because the max levels are inlined, validating both the upper and lower
bound of the level values is cheap enough to do, so we change to
validating these always, rather than just as a FILE_CHECK.

Refactor level decoding into LevelDecoder class so that we can use
internal BitReader in RleDecoder.

Eliminate dict_decoder_base_ member in favour of virtual function
since it is not called on hot paths.

Include DictDecoder inline to avoid pointer chasing in ReadSlot.

Avoid updating bitmap filter count when bitmap filter is disabled.

Cache the offsets from slot_desc_ inline in the struct.

Change-Id: I3fbabc770cb1578906f8605d02fd243d5d3747c8
Reviewed-on: http://gerrit.cloudera.org:8080/1122
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Tim Armstrong <tarmstrong@cloudera.com>