IMPALA-2517: bad buffer/block mapping in BufferedBlockMgr
The bug was that a recent change didn't set the BufferDesc to Block
pointer to NULL when deleting certain I/O blocks. This showed up as
a DCHECK in stress tests runs in a later call into BufferedBlockMgr once
the block had been recycled.

The issue can be detected earlier and reproduced with a simple unit test
by adding an additional DCHECK to BufferedBlockMgr::Validate() to check
that the BufferDesc<->Block relationship is consistent.

Change-Id: Ic93b22166de2eaf17bccb165ef7ff669f01c6897
Reviewed-on: http://gerrit.cloudera.org:8080/1219
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins