IMPALA-2491: remove TypeDescToColumnType() calls in hot functions
ColumnTypes are not cheap to construct, especially since adding the
extra field_names vector. TypeDescToColumnType() returns a ColumnType
and is used in many per-value functions. This patch removes these uses
and uses the FunctionContext::TypeDescs directly instead.

Running a local benchmark similar to that described in the JIRA, the
query time went from 5.37s to 2.98s with this change.

Change-Id: Icc201af68f9e7f5fa32a053b6343aa8c95bc75cb
Reviewed-on: http://gerrit.cloudera.org:8080/1227
Reviewed-by: Skye Wanderman-Milne <skye@cloudera.com>
Tested-by: Internal Jenkins