IMPALA-2103: Fix flaky test_impersonation test
This test would occasionally fail when an attempt to read a query
profile would occasionally return the empty string. The root cause was
that the Impalad that the query was submitted to, and the one used to
retrieve the profile from, could be different processes.

This patch changes the test to use the same process for HS2 and
webserver connections, by making the HS2 port available through the
ImpaladService object that was already used to access the webserver.

Previously the test would fail within ~30 minutes; with this patch it
has run for ~60 minutes without failure.

This patch also adds '--abort_on_failed_audit_event=false' to
test_impersonation to aid debugging.

Change-Id: Ib385438db96ee575d2f135c8b6459506b865e2d3
Reviewed-on: http://gerrit.cloudera.org:8080/1267
Reviewed-by: Henry Robinson <henry@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 7fc28793ed9f9dba6f94b2109ca12395ca21d557)
Reviewed-on: http://gerrit.cloudera.org:8080/1739
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Reviewed-by: Juan Yu <jyu@cloudera.com>
Tested-by: Juan Yu <jyu@cloudera.com>