IMPALA-2591: DataStreamSender::Send() does not return an error status if SendBatch() failed
If SendBatch() failed in DataStreamSender::Send(), we did not return
an error. This has been changed to accomodate the same.

A simple test case to test this is not straightforward to add and so I
haven't added any tests.

Change-Id: I4a63ba41a1b5b928f03bc0be1d178ce255825f9f
Reviewed-on: http://gerrit.cloudera.org:8080/1332
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 6af6cd7394109fbe96ef4864fa0daeefc3710d19)