Link Cyrus SASL dynamically in Impala
Until now, we would statically link SASL in Impala and simply hope
that it is ABI compatible to the version where we install
Impala. However, SASL is as well a security library and thus Impala
should not statically link it but rather implicitly depend on it.

The only complication of this patch is that there was an API breaking
change in SASL that this patch has to deal with as it compiles on
different platforms.

I tested this patch on Ubuntu and CentOS 6 with the two different
versions and everything worked fine.

Change-Id: I1b0925693f5fb20e9ce640c95f63d29234dc4f7c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/9150
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins