IMPALA-2632,IMPALA-799: improve LLVM IR building
Change build files to generate and compile C++ file with LLVM IR as C
array. The LLVM IR files are now generated as a dependency of the
codegen module, rather than separately. The IR files are kept in the
same location as before, but are not required at runtime.

Also change suffix from .ll (which is meant to be for the human-readable
text IR) to .bc (the correct extension for LLVM bitcode).

Includes a modified script file2array.sh from the Apache-licensed
Swift/T project that generates a valid C++ source file with binary
data encoded as an array variable.

Also enable and fix CMake rules to generate the IR metadata, instead of
doing it in make_impala.sh.

Change-Id: I432f5aa42f88b614af3339a465893c557192bbd5
Reviewed-on: http://gerrit.cloudera.org:8080/1387
Tested-by: Internal Jenkins
Reviewed-by: Martin Grund <mgrund@cloudera.com>