IMPALA-2648/IMPALA-2664: Avoid sending large partition stats objects over thrift
This patch sets a maximum limit of 200 megabytes on partition intermediate
stats that can be included in a THdfsTable object. We add this limit to
prevent the catalog jvm from hitting a max array limit error (and OOM) while
building the thrift objects to send to impalad. With this commit, an error is
thrown during the analysis of a COMPUTE INCREMENTAL STATS statement if the
same limit is hit based on the number of columns and the number of partitions
of a table (IMPALA-2664).

Change-Id: Ie420bc454ec19336110d5662b77182b9213cb388
Reviewed-on: http://gerrit.cloudera.org:8080/1425
Reviewed-by: Bharath Vissapragada <bharathv@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 697cd78ba61b1c32cdd9cc78c35d83da02738c79)
Reviewed-on: http://gerrit.cloudera.org:8080/1646
Reviewed-by: Juan Yu <jyu@cloudera.com>
Tested-by: Juan Yu <jyu@cloudera.com>
(cherry picked from commit 91d7674b508abc38fb5744558f68b65dca43e99f)