IMPALA-1656: Get rid of HashTable(MemPool*) constructor and add a TestEnv class
Until now some of the BE tests, especially the hash-table-test, were
using only a part of the runtime components Impalad is using. They
were not using ExecEnv, BufferBlockMgr, etc... That was leading to
testing codepaths that were slightly different that those exercized by
Impalad. The existence of those two different setups was also adding
some overhead as we were checking, for example, whether the
buffered-block-mgr is set or not (in Impalad it is always set).

This patch adds a new TestEnv class that contains most of the
components used at runtime (ExecEnv, TmpFileMgr, trackers) and
can create RuntimeStates with associated BufferedBlockMgrs to
simulate the state of query. This enables additional test coverage.
For example we can test the case where we don't have memory to
allocate even the bucket array (see IMPALA-1836).  Or the case where
we create a hash table with a very large number of buckets (>2^31)
that creates an arithmetic overflow (see IMPALA-2065).

If there are other tests of components that at Impalad use memory
coming from block-mgr we should convert them to use this new class as
well.

This patch removes the unnecessary codepaths in hash-table that were
dealing with the duality in the configurations of BE tests and
Impalad (sometimes the memory was being taken from the
buffered-block-mgr, others from a MemPool).

This patch changes the tests to set up the environment correctly with
at most one BufferedBlockMgr per RuntimeState.

Remove old function definition in SortNode().

Simplify BufferedBlockMgr by requiring that a MemTracker is provided for
each Client. Previously it was inconsistently required - e.g.
ConsumeMemory did not check for the client's MemTracker being NULL.

Original version of patch authored by Ippokratis Pandis
<ipandis@cloudera.com>

Change-Id: I4dff01414b1a7f7b8054caed0e81a885b846321c
Reviewed-on: http://gerrit.cloudera.org:8080/1561
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins