IMPALA-2722: Free local allocations per row batch in non-partitioned AGG and HJ
The fix of IMPALA-2612 added a test which verifies that the memory
limit is not exceeded for the given query if local allocations are
freed per processed row batch. The test fails when PHJ is disabled
as the same bug (of missing free calls to local allocations) also
exists in the non-partitioned counterpart of PAGG and PHJ. This
change fixes both non-partitioned AGG and HJ nodes so local
allocations are freed per row batch.

Change-Id: I9aec4d1fb4f8d52fcdf67f1ca545efa698764f84
Reviewed-on: http://gerrit.cloudera.org:8080/1569
Reviewed-by: Matthew Jacobs <mj@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Reviewed-by: Marcel Kornacker <marcel@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit ff75c9cde1f6554e7d7cef0b06af366eeabc0a1a)
Reviewed-on: http://gerrit.cloudera.org:8080/1824
Reviewed-by: Juan Yu <jyu@cloudera.com>
Tested-by: Huaisi Xu <hxu@cloudera.com>