IMPALA-2731: Refactor MemPool usage in HBase scan node.
Problem: The HBase scan node had an own MemPool that would regularly
give up the memory to the row batch. However, this would always pass in
true for the parameter that indicates if the last chunk should be kept,
except for the last batch, leaving the MemPool with exactly one very
large chunk that is partially allocated and continues to grow. Until
IMPALA-2284, at some point the chunk size would overflow and the MemPool
would fix this overflow accidentally by using a `max(req_size,
chunk_size)`, where chunk_size would be negative. After fixing
IMPALA-2284, the overflow no longer occurs, allowing for a rapid growth
of memory usage if the MemPool retains the last page.

The fix is that the HBase scan node now uses the MemPool of the row
batch to satisfy the allocations. The impact on the peak memory
consumption of the fragment is as follows:

  - Before IMPALA-2284: 2.8 GB
  - With IMPALA-2284: 23.06 GB
  - With IMPALA-2731(this): 5.79 MB

Change-Id: I9b499d83e221b0a58ff24bc32a212ec2e404c5c3
Reviewed-on: http://gerrit.cloudera.org:8080/1582
Reviewed-by: Dan Hecht <dhecht@cloudera.com>
Reviewed-by: Tim Armstrong <tarmstrong@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit d9de2ff89d2e7c06188be8cfde639bccdd2cc3c0)
Reviewed-on: http://gerrit.cloudera.org:8080/1643
Reviewed-by: Bharath Vissapragada <bharathv@cloudera.com>
Tested-by: Juan Yu <jyu@cloudera.com>