IMPALA-1170: Fix URL parsing when path contains '@'
RFC 3986 allows '@' characters in URLs to appear as the delimiter
between userinfo and host, as well as inside the path and in the
query. This change enables proper handling of an '@' character appearing
in either place or in all of them.

Change-Id: Ia4799d2dc25d2d840851938d3bb102e6f840b21c
Reviewed-on: http://gerrit.cloudera.org:8080/1584
Tested-by: Internal Jenkins
Reviewed-by: Alex Behm <alex.behm@cloudera.com>
(cherry picked from commit 01ee54f75625e566a412ec395db07b95296b9ea9)