IMPALA-2747: Thrift-client cleans openSSL state before using it in the case of the catalog
The TSSLSocketFactory was created with a local scope in the
thrift-client. This caused the factory to be destroyed on function
exit in the case of the catalog (see JIRA description for why it
happens only in the case of the catalog), which cleaned up the openSSL
library state set up by the thrift-client because Thrift
reference-counts shared TSSLSocketFactory state, and when that count
hits 0 OpenSSL is shut down. Once the OpenSSL library state was
cleaned up this way, it could not refer the trusted CA
(Certificate Authority) which the thrift client loaded. This led to
the client never being able to verify the server's certificate against
the CA.

This patch ties the lifetime of the TSSLSocketFactory to the socket
that relies on it to avoid this bug.

Change-Id: I54b9e7bf8d1766fb5154f3a826c0297ee99eb87e
Reviewed-on: http://gerrit.cloudera.org:8080/1594
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Internal Jenkins
(cherry picked from commit 79e484f3c2771d3b5949cc385e86238fe48f250c)