IMPALA-2740: Kudu scanner - Reset tuple's null bits after filtering row
The problem was, if a tuple was filtered, the bits indicating values are
NULL were not reset and the tuple's memory was reused. So NULLs from
consecutively filtered rows would accumulate in the tuple. The fix is to
always reset the NULL bits (as it doesn't matter whether the row was
filtered).

Change-Id: Ib4d980980e02bf2c82dc229a8ed1ada16bb8174f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/9958
Tested-by: jenkins
Reviewed-by: Martin Grund <mgrund@cloudera.com>